requirements.txt
# Core LangGraph and LangChain
langgraph==0.2.28
langchain==0.3.3
langchain-core>=0.3.15,<0.4
langchain-community==0.3.2

# LLM Providers
langchain-openai==0.2.1
langchain-anthropic==0.2.3
langchain-google-genai==2.0.4

# Azure OpenAI
openai==1.51.2

# API Clients
anthropic==0.39.0
google-generativeai==0.8.3
requests==2.32.3

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Data Processing
httpx==0.27.2

# Logging and Monitoring
loguru==0.7.2

# Output Formatting
tabulate==0.9.0
markdown==3.7

# JSON Processing
jsonschema==4.23.0

# Utility
python-dateutil==2.9.0.post0

# Web Framework
flask==3.1.0

# PDF Generation
reportlab==4.0.7
markdown2==2.4.12
